requirements.txt
flask==2.3.3
web3==6.11.3
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
flask-cors==4.0.0
